requirements.txt
torch
torchvision
torchaudio
numpy
pandas
matplotlib
opencv-python
scikit-learn
pillow
tqdm
jupyter